requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12
Flask-Login==0.4.0
Flask-SQLAlchemy==2.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
mysql-connector==2.2.3
numpy==1.13.0
pygame===1.9.1release
rauth==0.7.3
requests==2.18.4
SQLAlchemy==1.1.14
urllib3==1.22
virtualenv==15.1.0
Werkzeug==0.11.15
google-api-python-client == 1.5.2
oauth2client==2.2.0
simplekv==0.10.0
flask_kvsession